docs/requirements.txt
mkdocs-material==9.7.0
mkdocstrings==0.30.1
mdx-truly-sane-lists==1.3
mkdocs-awesome-pages-plugin==2.10.1
mkdocstrings==1.17.0
pymarkdownlnt==0.9.33
setuptools==80.9.0
